+++ internal/authz/mock/noop_authz.go
@@ -14,44 +14,41 @@
 // limitations under the License.
 
 // Package fake provides a no-op implementation of the minder the authorization client
-package fake
+package mock
 
 import (
 	"context"
-	"slices"
+	"fmt"
 
 	"github.com/google/uuid"
 
 	"github.com/stacklok/minder/internal/authz"
 )
 
-// SimpleClient maintains a list of authorized projects, suitable for use in tests.
-type SimpleClient struct {
-	Allowed []uuid.UUID
+// NoopClient is a no-op implementation of the authz.Client interface, which always returns
+// the same authorization result.
+type NoopClient struct {
+	// If Authorized is true, all Check calls will return nil
+	Authorized bool
 }
 
-var _ authz.Client = &SimpleClient{}
+var _ authz.Client = &NoopClient{}
 
 // Check implements authz.Client
-func (n *SimpleClient) Check(_ context.Context, _ string, project uuid.UUID) error {
-	if slices.Contains(n.Allowed, project) {
+func (n *NoopClient) Check(_ context.Context, action string, project uuid.UUID) error {
+	fmt.Printf("noop authz check (%t): %s %s\n", n.Authorized, action, project)
+	if n.Authorized {
 		return nil
 	}
 	return authz.NotAuthorized
 }
 
-// Write implements authz.Client
-func (n *SimpleClient) Write(_ context.Context, _ string, _ authz.AuthzRole, project uuid.UUID) error {
-	n.Allowed = append(n.Allowed, project)
+// Write_ implements authz.Client
+func (_ *NoopClient) Write(_ context.Context, _ string, _ authz.AuthzRole, _ uuid.UUID) error {
 	return nil
 }
 
 // Delete implements authz.Client
-func (n *SimpleClient) Delete(_ context.Context, _ string, _ authz.AuthzRole, project uuid.UUID) error {
-	index := slices.Index(n.Allowed, project)
-	if index != -1 {
-		n.Allowed[index] = n.Allowed[len(n.Allowed)-1]
-		n.Allowed = n.Allowed[:len(n.Allowed)-1]
-	}
+func (_ *NoopClient) Delete(_ context.Context, _ string, _ authz.AuthzRole, _ uuid.UUID) error {
 	return nil
 }
